requirements.txt
tweepy
Image
numpy
imutils
contours
scikit-image
opencv-python
twitter-dl
argparse